factor lookup into file
