+++ dogsdogsdogs/examples/ngo.rs
@@ -68,15 +68,17 @@ where G::Timestamp: Lattice+Hash+Ord {
     // determine for each (src, dst) tuple which index would propose the fewest extensions.
     let winners = cand_count1.concat(&cand_count2)
                              .reduce(|_srcdst, counts, output| {
-                                 let mut min_cnt = isize::max_value();
-                                 let mut min_idx = usize::max_value();
-                                 for &(&idx, cnt) in counts.iter() {
-                                     if min_cnt > cnt {
-                                         min_idx = idx;
-                                         min_cnt = cnt;
+                                 if counts.len() == 2 {
+                                     let mut min_cnt = isize::max_value();
+                                     let mut min_idx = usize::max_value();
+                                     for &(&idx, cnt) in counts.iter() {
+                                         if min_cnt > cnt {
+                                             min_idx = idx;
+                                             min_cnt = cnt;
+                                         }
                                      }
+                                     output.push((min_idx, 1));
                                  }
-                                 output.push((min_idx, 1));
                              });
 
     // select tuples with the first relation minimizing the proposals, join, then intersect.
