+++ dogsdogsdogs/src/operators/mod.rs
@@ -1,8 +1,10 @@
+pub mod lookup_map;
+
 pub mod count;
 pub mod propose;
 pub mod validate;
 
-pub use self::propose::lookup_map;
+pub use self::lookup_map::lookup_map;
 pub use self::count::count;
-pub use self::propose::propose;
+pub use self::propose::{propose, propose_distinct};
 pub use self::validate::validate;
